Add user and tag policies and use them in the controllers

+++ app/Http/Controllers/TagController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Tag;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class TagController extends Controller
 {
@@ -79,6 +80,8 @@ class TagController extends Controller
      */
     public function edit(Tag $tag)
     {
+        abort_unless(Gate::allows('update', $tag), 403);
+
         return view('tag.edit')->with([
             'tag' => $tag
         ]);
@@ -93,6 +96,8 @@ class TagController extends Controller
      */
     public function update(Request $request, Tag $tag)
     {
+        abort_unless(Gate::allows('update', $tag), 403);
+
         $request->validate([
             'name' => 'required|min:3',
             'style' => 'required'
@@ -115,6 +120,8 @@ class TagController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        abort_unless(Gate::allows('delete', $tag), 403);
+
         $oldName = $tag->name;
         $tag->delete();
         return $this->index()->with([
